lab 10 feito

+++ src/main/java/pt/ulusofona/cd/store/service/SupplierService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pt.ulusofona.cd.store.client.OrderClient;
 import pt.ulusofona.cd.store.client.ProductClient;
+import pt.ulusofona.cd.store.client.ProductGrpcClient;
 import pt.ulusofona.cd.store.dto.*;
 import pt.ulusofona.cd.store.events.SupplierEventProducer;
 import pt.ulusofona.cd.store.exception.SupplierNotFoundException;
@@ -24,6 +25,7 @@ public class SupplierService {
     private final ProductClient productClient;
     private final OrderClient orderClient;
     private final SupplierEventProducer supplierEventProducer;
+    private final ProductGrpcClient productGrpcClient;
 
     @Transactional
     public Supplier createSupplier(SupplierRequest request) {
@@ -106,6 +108,14 @@ public class SupplierService {
     @Transactional
     public void deleteSupplier(UUID id) {
         Supplier supplier = getSupplierById(id);
+
+        int count = productGrpcClient.countDiscontinuedProducts(id.toString(), false);
+        if (count > 0) {
+            throw new IllegalStateException(
+                    "Supplier cannot be deleted because they still have " + count + " active products associated."
+            );
+        }
+
         supplierRepository.delete(supplier);
     }
